+++ src/tests/orga/views/test_orga_views_speaker.py
@@ -1,3 +1,5 @@
+import json
+
 import pytest
 from django_scopes import scope, scopes_disabled
 
@@ -262,3 +264,94 @@ def test_orga_can_delete_speaker_information(orga_client, event, information):
     orga_client.post(information.orga_urls.delete, follow=True)
     with scope(event=event):
         assert event.information.all().count() == 0
+
+
+@pytest.mark.django_db
+def test_orga_cant_export_answers_csv_empty(orga_client, speaker, event, submission):
+    response = orga_client.post(
+        event.orga_urls.speakers + "export/",
+        data={
+            "target": "rejected",
+            "name": "on",
+            "export_format": "csv",
+        },
+    )
+    assert response.status_code == 200
+    assert (
+        response.content.decode().strip().startswith("<!DOCTYPE")
+    )  # HTML response instead of empty download
+
+
+@pytest.mark.django_db
+def test_orga_cant_export_answers_csv_without_delimiter(
+    orga_client, speaker, event, submission, answered_choice_question
+):
+    with scope(event=event):
+        answered_choice_question.target = "speaker"
+        answered_choice_question.save()
+    response = orga_client.post(
+        event.orga_urls.speakers + "export/",
+        data={
+            "target": "all",
+            "name": "on",
+            f"question_{answered_choice_question.id}": "on",
+            "export_format": "csv",
+        },
+    )
+    assert response.status_code == 200
+    assert response.content.decode().strip().startswith("<!DOCTYPE")
+
+
+@pytest.mark.django_db
+def test_orga_can_export_answers_csv(
+    orga_client, speaker, event, submission, answered_choice_question
+):
+    with scope(event=event):
+        answered_choice_question.target = "speaker"
+        answered_choice_question.save()
+        answer = answered_choice_question.answers.all().first().answer_string
+    response = orga_client.post(
+        event.orga_urls.speakers + "export/",
+        data={
+            "target": "all",
+            "name": "on",
+            f"question_{answered_choice_question.id}": "on",
+            "submission_ids": "on",
+            "export_format": "csv",
+            "data_delimiter": "comma",
+        },
+    )
+    assert response.status_code == 200
+    assert (
+        response.content.decode()
+        == f"ID,Name,Proposal IDs,{answered_choice_question.question}\r\n{speaker.code},{speaker.name},{submission.code},{answer}\r\n"
+    )
+
+
+@pytest.mark.django_db
+def test_orga_can_export_answers_json(
+    orga_client, speaker, event, submission, answered_choice_question
+):
+    with scope(event=event):
+        answered_choice_question.target = "speaker"
+        answered_choice_question.save()
+        answer = answered_choice_question.answers.all().first().answer_string
+    response = orga_client.post(
+        event.orga_urls.speakers + "export/",
+        data={
+            "target": "all",
+            "name": "on",
+            f"question_{answered_choice_question.id}": "on",
+            "submission_ids": "on",
+            "export_format": "json",
+        },
+    )
+    assert response.status_code == 200
+    assert json.loads(response.content.decode()) == [
+        {
+            "ID": speaker.code,
+            "Name": speaker.name,
+            answered_choice_question.question: answer,
+            "Proposal IDs": [submission.code],
+        }
+    ]