+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`313` Organisers can now create custom speaker exports (as either CSV or JSON), including any speaker data they need.
 - :support:`-` If you use PostgreSQL, please make sure to be on at least PostgreSQL 10.
 - :feature:`-` pretalx will no longer send emails to localhost addresses, as those are used for deleted users. Please get in touch if this is a problem for you.
 - :feature:`1069` You can freeze a question after a certain date, prohibiting users from changing their answers after the deadline.